--- src/main/java/nextstep/subway/applicaion/error/ErrorResponse.java
@@ -1,20 +1,0 @@
-package nextstep.subway.applicaion.error;
-
-public class ErrorResponse {
-	private String message;
-
-	private ErrorResponse() {
-	}
-
-	private ErrorResponse(String message) {
-		this.message = message;
-	}
-
-	public static ErrorResponse from(String message) {
-		return new ErrorResponse(message);
-	}
-
-	public String getMessage() {
-		return message;
-	}
-}
