--- src/main/java/nextstep/subway/applicaion/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.applicaion.exception;
-
-public abstract class BadRequestException extends RuntimeException {
-
-	public BadRequestException(String message) {
-		super(message);
-	}
-}
