--- src/main/java/nextstep/subway/applicaion/exception/ConflictRequestException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.applicaion.exception;
-
-public abstract class ConflictRequestException extends RuntimeException {
-
-	public ConflictRequestException(String message) {
-		super(message);
-	}
-}
