--- src/main/java/nextstep/subway/applicaion/exception/EntityNotFoundException.java
@@ -1,10 +1,0 @@
-package nextstep.subway.applicaion.exception;
-
-import static java.lang.String.format;
-
-public class EntityNotFoundException extends BadRequestException {
-
-	public EntityNotFoundException(long id) {
-		super(format("Entity key '%s' is not exists", id));
-	}
-}
