--- src/main/java/nextstep/subway/applicaion/exception/UniqueKeyExistsException.java
@@ -1,10 +1,0 @@
-package nextstep.subway.applicaion.exception;
-
-import static java.lang.String.format;
-
-public class UniqueKeyExistsException extends ConflictRequestException {
-
-	public UniqueKeyExistsException(String existsUniqueKey) {
-		super(format("Unique Key '%s' is exists", existsUniqueKey));
-	}
-}
