--- src/main/java/nextstep/subway/applicaion/station/repository/StationRepository.java
@@ -1,14 +1,0 @@
-package nextstep.subway.applicaion.station.repository;
-
-import nextstep.subway.applicaion.station.domain.Station;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-	@Override
-	List<Station> findAll();
-
-	Optional<Station> findByName(String name);
-}
