--- src/main/java/nextstep/subway/applicaion/station/service/StationQueryService.java
@@ -1,23 +1,0 @@
-package nextstep.subway.applicaion.station.service;
-
-import nextstep.subway.applicaion.station.repository.StationRepository;
-import nextstep.subway.applicaion.station.response.StationResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@Transactional(readOnly = true)
-public class StationQueryService {
-	private final StationRepository stationRepository;
-
-	public StationQueryService(StationRepository stationRepository) {
-		this.stationRepository = stationRepository;
-	}
-
-	public List<StationResponse> findAllStations() {
-		return StationResponse.fromList(stationRepository.findAll());
-	}
-
-}
