+++ src/main/java/nextstep/subway/application/dto/request/LineRequest.java
@@ -1,0 +1,53 @@
+package nextstep.subway.application.dto.request;
+
+import nextstep.subway.domain.Line;
+
+import java.util.Objects;
+
+public class LineRequest {
+	private String name;
+	private String color;
+	private Long upStationId;
+	private Long downStationId;
+	private int distance;
+
+	private LineRequest() {
+	}
+
+	public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+		this.name = name;
+		this.color = color;
+		this.upStationId = upStationId;
+		this.downStationId = downStationId;
+		this.distance = distance;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public String getColor() {
+		return color;
+	}
+
+	public Line toEntity() {
+		return Line.of(name, color);
+	}
+
+	public Long getUpStationId() {
+		return upStationId;
+	}
+
+	public Long getDownStationId() {
+		return downStationId;
+	}
+
+	public int getDistance() {
+		return distance;
+	}
+
+	public boolean isStationNotProvided() {
+		return Objects.isNull(upStationId) ||
+				Objects.isNull(downStationId);
+	}
+}
