+++ src/main/java/nextstep/subway/application/exception/BadRequestException.java
@@ -1,0 +1,8 @@
+package nextstep.subway.application.exception;
+
+public abstract class BadRequestException extends RuntimeException {
+
+	public BadRequestException(String message) {
+		super(message);
+	}
+}
