+++ src/main/java/nextstep/subway/application/exception/ConflictRequestException.java
@@ -1,0 +1,8 @@
+package nextstep.subway.application.exception;
+
+public abstract class ConflictRequestException extends RuntimeException {
+
+	public ConflictRequestException(String message) {
+		super(message);
+	}
+}
