+++ src/main/java/nextstep/subway/application/exception/UniqueKeyExistsException.java
@@ -1,0 +1,10 @@
+package nextstep.subway.application.exception;
+
+import static java.lang.String.format;
+
+public class UniqueKeyExistsException extends ConflictRequestException {
+
+	public UniqueKeyExistsException(String existsUniqueKey) {
+		super(format("Unique Key '%s' is exists", existsUniqueKey));
+	}
+}
