+++ src/main/java/nextstep/subway/application/repository/StationRepository.java
@@ -1,0 +1,14 @@
+package nextstep.subway.application.repository;
+
+import nextstep.subway.domain.Station;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface StationRepository extends JpaRepository<Station, Long> {
+	@Override
+	List<Station> findAll();
+
+	Optional<Station> findByName(String name);
+}
