+++ src/main/java/nextstep/subway/application/service/StationService.java
@@ -1,0 +1,38 @@
+package nextstep.subway.application.service;
+
+import nextstep.subway.application.dto.request.StationRequest;
+import nextstep.subway.application.dto.response.StationResponse;
+import nextstep.subway.application.exception.UniqueKeyExistsException;
+import nextstep.subway.application.repository.StationRepository;
+import nextstep.subway.domain.Station;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class StationService {
+	private final StationRepository stationRepository;
+
+	public StationService(StationRepository stationRepository) {
+		this.stationRepository = stationRepository;
+	}
+
+	@Transactional
+	public StationResponse saveStation(StationRequest requestStation) {
+		if (isStationExists(requestStation)) {
+			throw new UniqueKeyExistsException(requestStation.getName());
+		}
+
+		final Station station = stationRepository.save(requestStation.toEntity());
+		return StationResponse.from(station);
+	}
+
+	@Transactional
+	public void deleteStationById(Long id) {
+		stationRepository.deleteById(id);
+	}
+
+	private boolean isStationExists(StationRequest station) {
+		return stationRepository.findByName(station.getName()).isPresent();
+	}
+}
